Workers can now be used via the normal programmatic API

+++ src/workers/Worker.js
@@ -164,8 +164,10 @@ class Worker {
     }
   }
   kill() {
-    this.logger(`killing worker process ${this.pid}`);
-    this.worker.kill();
+    if (this.worker.connected) {
+      this.logger(`killing worker process ${this.pid}`);
+      this.worker.kill();
+    }
   }
 }
